+++ R/plot.desc.R
@@ -61,8 +61,7 @@
 #' 
 #'test3=emmeans(mod3,~GROUP|TIMEPOINT)
 #' 
-#'tab.mod3=report.lsmeans(lsm=test3,x1="GROUP",
-#'		x2="TIMEPOINT",at.row="TIMEPOINT",data=data.mod)
+#'tab.mod3=report.lsmeans(lsm=test3,at.row="TIMEPOINT")
 #' 
 #'gg=plot(tab.mod3,title="LS Means plot example")
 #'