+++ R/time_to_cure-data.R
@@ -1,0 +1,22 @@
+# TODO: Add comment
+# 
+# Author: jfcollin
+###############################################################################
+
+
+#' Fake clinical data example
+#'
+#' Fictive data created only for the purpose of testing the package and showing examples.
+#'
+#' @format A data frame
+#' 
+#' @docType data
+#'
+#' @source From pure imagination
+#' 
+#' @keywords datasets
+#' 
+#' @examples
+#' data(time_to_cure)
+#' \donttest{head(time_to_cure)}
+"time_to_cure"